cleanup for omnilang integration
