+++ omniplanner/src/dsg_pddl/pddl_grounding.py
@@ -120,7 +120,7 @@ class PddlDomain:
         self.domain_ast = lisp_string_to_ast(domain_str)
         self.domain_name = get_domain_name(self.domain_ast)
         self.requirements = get_domain_requirements(self.domain_ast)
-        self.get_domain_types = get_domain_types(self.domain_ast)
+        self.domain_types = get_domain_types(self.domain_ast)
         self.predicates = get_domain_predicates(self.domain_ast)
         self.functions = get_functions(self.domain_ast)
         self.derived = get_derived(self.domain_ast)
